Allow middle row/column to have zero height.
The code appears to be working on the assumption that there will always be content, thus "auto" makes sense, but it breaks pretty badly when there is no content and the height is supposed to match the parent.

Combined with the previous change to allow width/height mutation, this provides an escape hatch that the app can use to get proper fit-to-parent height (sort of) in the middle row.